Update README.md
https://github.com/krish08jack/SPARK-FOUNDATION-ASSIGNMENT-1.git
Linear regression is a statistical model that analyzes the relationship between a response variable (often called y) and one or more variables and their interactions (often called x or explanatory variables).
The summary outputs show 6 components, including:
Call. Shows the function call used to compute the regression model.
Residuals. Provide a quick view of the distribution of the residuals, which by definition have a mean zero. Therefore, the median should not be far from zero, and the minimum and maximum should be roughly equal in absolute value.
Coefficients. Shows the regression beta coefficients and their statistical significance. Predictor variables, that are significantly associated with the outcome variable, are marked by stars.
The residual standard error (RSE), R-squared (R2), and the F-statistic are metrics that are used to check how well the model fits our data.
The coefficients table, in the model statistical summary, shows:

the estimates of the beta coefficients
the standard errors (SE), which defines the accuracy of beta coefficients. For a given beta coefficient, the SE reflects how the coefficient varies under repeated sampling. It can be used to compute the confidence intervals and the t-statistic.
the t-statistic and the associated p-value, which defines the statistical significance of the beta coefficients.